requirements/requirements.txt
aerich==0.7.2
aiohttp==3.11.8
asyncmy==0.2.9
boto3==1.35.26
orjson==3.10.7
sanic==24.6.0
sanic-ext==23.12.0
tortoise-orm==0.21.7
redis==5.1.1
